Add XML.AbortOnErrors and XML.AbortOnWarnings
It is sometimes useful for warnings or errors generated by 
the FoX libary to result in an abort to generate a core dump or 
stack trace to see where in Siesta the call causing the error 
is being made. It is now possible to turn on this behaviour at 
runtime using two new fdf commands:

XML.AbortOnErrors
XML.AbortOnWarnings

Both are boolean and default to .false. (which mirrors the 
current stopping behaviour).