Update FoX library and add error and warnings handling options (A. Walker)
It is sometimes useful for warnings or errors generated by the FoX
libary to result in an abort to generate a core dump or stack trace to
see where in Siesta the call causing the error is being made. It is
now possible to turn on this behaviour at runtime using two new fdf
boolean tags:

 XML.AbortOnErrors  XML.AbortOnWarnings

Both are boolean and default to .false. (which mirrors the current
stopping behaviour).

These features have been made possible by an update in the FoX library
that adds an interface to allow errors and warnings to result in an
abort signal. In the update there is also a fix to work around a PGI
bug for the overloaded wxml interfaces (aparently not used in Siesta)
and a change to the prng to avoid using initalised data in prng_init.

Documentation for these new options is in the XML appendix of
siesta.tex

The as-yet undocumented WriteXML option has been renamed to XML.Write
for consistency, although the old form is retained for now.

added:
  Src/FoX/arch.make
modified:
  Docs/siesta.tex
  Src/FoX/Makefile
  Src/FoX/common/FoX_common.F90
  Src/FoX/common/m_common_error.F90
  Src/FoX/common/makefile
  Src/FoX/fsys/fox_m_fsys_abort_flush.F90
  Src/FoX/utils/fox_m_utils_uuid.F90
  Src/FoX/utils/makefile
  Src/FoX/wcml/m_wcml_molecule.F90
  Src/FoX/wcml/makefile
  Src/FoX/wxml/m_wxml_core.F90
  Src/FoX/wxml/m_wxml_overloads.F90
  Src/FoX/wxml/makefile
  Src/siesta_cml.f90
  Src/siesta_cmlsubs.F90