Avoid extrapolation of DM during geometry optimization
When performing a geometry optimization (CG, Broyden, etc) (idyn=0 in
the code), there might be sudden jumps in the geometry (e.g., at the
end of a line-minimization) that render a extrapolation of the DM
risky. Hence DM extrapolation is turned off in this case.

One might argue that DM re-use should be avoided too.

modified:
  Docs/CHANGES
  Docs/siesta.tex
  Src/new_dm.F