Fix bug in previous patch regarding CG DM extrapolation
The logic of a previous patch was wrong, and the resulting
code prevented the proper restructuring of a reused DM in
the course of CG relaxations.

To make it clear now: For "CG" calculations, the default is not to
extrapolate the DM (unless requested by explicitly setting
'DM.AllowExtrapolation' to "true"). The previous step's DM is reused.

modified:
  Docs/siesta.tex
  Src/new_dm.F