Fix bug in "kick" code in Pulay mixing
The program was not taking into account the requested value of the
DM.KickMixingWeight coefficient if the 'DM.NumberKick N' option was
specified in combination with Pulay mixing. It was always set
implicitly to 1.0 (i.e., every N steps the full "output" density
matrix was taken as the input for the next iteration, instead of the
appropriate mix of input and output DMs).

The bug has been fixed.

Modified:
  Src/m_pulay.F