Experimental output comparison tool
A simple "output digest" script in awk extracts key information fields from the output
file with default precisions, and generates a small file. These digest files can then
be compared to check for differences.

This is an experimental feature. It is not yet integrated in the automatic test
framework. Note that the reference output files in Tests/Reference have not yet
been updated, pending a review.

added:
  Tests/cmp_digest.sh               ! Script to compare digest files
  Tests/out_digest.awk              ! Digest generator
  Tests/out_digest_template.awk     ! Template for optional precision changes
modified:
  Src/obj_setup.sh                  ! It now creates a symbolic link to the ref directory.