Avoid reporting unstable inversion in pulayx
Typically, inversion problems in the DIIS procedure occur when the
error estimate for a step is very small compared to the previous
ones. This is harmless, and the reporting of a failure in the output
file is confusing.

In case it is needed, a hook to allow a full diagnosis 
can be compiled in with the preprocessor flag

DEBUG_PULAY_INVERSE

(+ Also, initialize "found" variable in new_dm.F in transiesta mode)

modified:
  Src/m_pulay.F90
  Src/new_dm.F