Put in safeguards for orbital indexing
When domain decomposition is used, the full indexing
arrays of the "spatial" mode (including a nL2G array
holding global information) are used in the LocalToGlobalOrb
and GetNodeOrbs calls. 

Calls to these routines in a non-local form is not allowed
in the initial stage, before the actual decomposition.

A. Garcia, May 2, 2011