Remove a superfluous test in new parallelsubs code
The version of parallelsubs used to support the domain (metis)
decomposition in the O(N) section included a test for null grid
that caused problems in some TranSiesta runs. This test seems
to be superfluous, and has been commented out. Some debugging
statements in m_ts_in_siesta.F have been kept (commented) to
help further investigation in case it is needed.

modified:
  Src/m_ts_in_siesta.F
  Src/parallelsubs.F