Implementation of Bader charge output
If the option

  SaveBaderCharge T

is used, the program will write the electronic charge density with an
extra set of model core charges, useful for post-processing by a Bader
"atoms in molecules" analysis. See the manual for important details.

The grid2cube program in Util has been extended to deal with the .BADER
file.

modified:
  Docs/siesta.ind
  Docs/siesta.tex
  Src/dhscf.F
  Src/meshsubs.F
  Src/siesta_options.F90
  Tests/Reference-xml/h2o-pop.xml
  Tests/Reference/h2o-pop.out
  Tests/h2o-pop/h2o-pop.fdf