Improved the estimation of the size of the automatic cell
If not specified, Siesta generates automatically a unit cell for
"molecular" or "atomic" systems. The criterion is that the cell is big
enough to avoid interactions between orbitals in two different cells.

Up to now, the estimation of the size of the orbitals was given by
the cutoff of Vna. But if the basis set contains large orbitals which
are not occupied in the ground state, this estimate is too short.

The "size of the atom" is now estimated as the largest of rc(Vna) and
the cutoffs of all the orbitals. 

modified:
  Src/Makefile
  Src/automatic_cell.f